pyproject.toml
[tool.poetry]
name = "whohosts"
version = "0.1.0"
description = ""
authors = ["rdpickard <code@chalkfarm.mx>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
requests = "^2.28.1"
Flask-RESTful = "^0.3.9"
redis = "^4.3.4"
Flask-Caching = "^2.0.1"
jsonschema = "^4.14.0"
netaddr = "^0.8.0"
dnspython = "^2.2.1"
arrow = "^1.2.2"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
